Refactor: Rename directory to make it even more consistent.
Refs: BLAIS5-3651
