+++ src/components/questionnaireDetailsPage/viewCatiModeDetails.test.tsx
