+++ src/components/questionnaireDetailsPage/viewCatiModeDetails.tsx
